requirements.txt
autopep8==1.5.7
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
pycodestyle==2.7.0
PyJWT==2.1.0
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
